Updated openstack/openstack Project: openstack/oslo.db  e734e2befb7c2e796c24106f3236d010a62e6b9b
Fix test to not assume eventlet isn't present

The test suite TpoolDbapiWrapperTestCase makes a blanket assumption
that the test environment does not have eventlet installed; it
places a mock object into sys.modules in order to take the place
of eventlet.   The test_db_api_without_installed_eventlet however
deletes this mock, assuming that subsequent access will raise an
ImportError.  The change is to instead place None in this dictionary
so that an ImportError is guaranteed regardless of whether or not
eventlet is actually installed in the environment.

Change-Id: I7b5b7c0d20581f8ecfa319f3cfa0f48112bc86b9
Closes-Bug: #1334427