Updated openstack/openstack Project: openstack/tempest  d08b9adac28d14cb1e7996f0e86ae8e284631f40
Handle backup not found in test_create_backup

The "oldest" backup in the test_create_backup test is supposed
to be deleted automatically via rotation but the cleanup routine
does not catch the NotFound exception when the delete function
doesn't find the backup image. This commit handles the NotFound
and also logs a warning if the image was actually deleted in
case of a race condition.

Change-Id: Ieae2edf0ae4318aef74aedef18480df165fd4753
Closes-Bug: #1291683