Updated openstack/openstack Project: openstack/oslo.messaging  0cfafac2468afe6b601a985495382aae3351b217
encoding error in file

There is a non-printable character (c2a0) in a comment line, which
in a new Neutron test case causes a StringException. There are
three issues here.

1) Encoding issue in the file.
2) No detection of encoding issues in files.
3) Neutron tests mask the failure detail

This addresses the first issue. Bug 1334798 in Neutron was created
for the last issue. Oslo team is aware of the second issue.

Related-Bug: #1334798

Change-Id: I8cfa6c93085a26666a8b4bee06dcd52709d04a95