Updated openstack/openstack Project: openstack/nova  598adddd65d181b81c25ad945e2c5b55a36dfcf5
Sync revert and finish resize on instance.uuid

We need to make sure we sync the state of the instance the same way we
do for do_confirm_resize. Previously, we were not syncing the instance
on revert or finish resize. We do this so that we can update the
migration and instance resize update orders. Previously if you set
resize_confirm_window to some small value, your resized instance could
go into error state because the periodic task was run after the
migration was updated but before the instance object was updated. This
change makes it so the instance is always updated before the migration.

Change-Id: Iaad1a05864b2b46180cadf16198b0fd8583c6703
Closes-Bug: 1326778