Updated openstack/openstack Project: openstack/nova  ae89f8b888c3fb2ee752a9c070af8694f22605f7
Fix update agent return agent_id with string for os-agents v3

Currently the agent update return agent_id with integer, that is
inconsistent with the response of agent create and index. It because
agent update use the agent_id that passed from url as the response.
This problem is fixed by converting the agent id as integer.

And the API sample tests provide agent_id with string type in the fake
data. This patch fix this problem for API sample files and tests.

And for back-compatibility, we should add translator for v2.1 API later.
Then this fix will be exposed after v3 API feature exposed by microversion
in the future.

DocImpact: The agent_is is changed to integer type for agents'update action.

Change-Id: Ia9c1034a325b77a2057934cc80b6eab679ff1d2b
Partial-Bug: #1333494