Updated openstack/openstack Project: openstack/cinder  67e32ac546f2c04e8e8dee2ad4e4dfd8b8f467a6
Misuse of i18n log translation to output error message

I found a misuse of i18n log translation for error message
in "cinder/volume/drivers/lvm.py'.

The "%s' can not be translated. In this case, the "message'
must be i18ning and then the "message" should be output
using LOG.error().

Change-Id: I4495522d337764dc28186225f4594a227ac3e555
Signed-off-by: Mitsuhiro Tanino <mitsuhiro.tanino@hds.com>
Closes-Bug: 1334334