Updated openstack/openstack Project: openstack/nova  afb914ee7f6e621f65d9f82608a30a74ab7e891c
Log when nova-conductor connection established

The current warning in wait_until_ready() is a bit of a red-herring.
When just looking at the logs, it's warning status may lead you to
believe the failure to connect to nova-conductor is fatal when really
it was retried and connected successfully.

The WARNING message on timeout is modified to clearly show that it is
retrying the connection.  A partner message is shown to inform that
the connection has been established.

Change-Id: I2040707df7caf8685c84c82ec9b8cc97cb04b47e