Updated openstack/openstack Project: openstack/cinder  be825174404c7867757c24ed5c83f8980b251278
Fixes cinder volume from snapshot on Windows

When creating a volume from a snapshot on Windows, a shadow copy
volume is exported as an iSCSI disk. The issue is that this export
is readonly and cannot be mounted to instances.

As this export cannot be modified, to make the new volume usable,
the according image must be moved to a new path and imported .

Closes-Bug: #1306032

Change-Id: I3a936d30fdd7875059dc56c2681f453757c2605f