Updated openstack/openstack Project: openstack/nova  9e61d37c19001110869e745037ecec7426f63e2b
VMware: Fix fake raising the wrong exception in _remove_file

fake was raising exception.FileNotFound error if a file was not found.
However, an actual delete on a non-existent file raises the error
returned by vSphere, which is error_util.FileNotFoundException. This
breaks tests which exercise error paths expecting the real error to be
raised.

Fortunately this change doesn't break any existing tests. However, I'm
not sure that's a good thing.

Change-Id: I2025bffa887582eaa9e9072d0400f90ca97d1898