Updated openstack/openstack Project: openstack/trove  76c2e6a88738fa53b3e7976b1adbfe95dde865ae
Add timestamps and instance count to config groups

This commit adds created and updated timestamps to the configuration
groups list and details calls. It also adds instance_count to the
configuration details call. Timestamps on configuration groups can
serve as a secondary indicator for making config edits, as well as a
potential time marker for correlating when detrimental config changes
were made. It will also help to see how many instances use a
particular config and users will be able to see up front how many
instances the config changes will affect.

Note that instance_count is generated on the fly and not stored in the
config db, hence the getattr() usage.

Change-Id: I0138a28e9fb4c0c045a1a26df3f8a282d80a7274
Implements: blueprint minor-config-edits
Docimpact: change to configuration payloads