Updated openstack/openstack Project: openstack/tempest  499aae3fcf35f0c0e0367c6898b018ca964254c5
Correct nc command in test_load_balancer_basic

HA-Proxy always expects it will close the connection first.
To make a backend respect it, the command that starts
the netcat "web server" should be modified by adding
'cat >/dev/null' to a responce. Thus a client is made to wait
until the proxy closes the connection.

Closes bug: 1326607

Change-Id: I762f7b59bd1fd76471cf5530882b6b08e8c9bb0b