Updated openstack/openstack Project: openstack/taskflow  70d5e694ff214149368318f2efc0c31763e5b46a
Clarify locked decorator is for instance methods

Make it clear when reading the lock utils code that the
locked decorator will only currently work with instance
methods by removing the usage of 'args[0]' and renaming
this variable to 'self' and adjusting the docstring.

Once the wrapt module is approved and we are able to use
it in taskflow, we can refactor the function to be correct
with regards to classmethods.

Change-Id: Ic33eb9e47679d2105654634469dd6d305d38b2e0