Updated openstack/openstack Project: openstack/horizon  56d71e22495b1afd464ead493ede04d8a9338fc7
Copy the selected file name in the object name field.

In the containers panel, when uploading a object, if the user
selects a file and the object name field is empty, copy the
file name in the object name field. The filename string is
saved in the file input as a attribute for comparison of any
future changes to the file input.

Moved the file input to improve the UX

Change-Id: I608ebd572a745780fb59af852b182d75ab300ce7
Closes-Bug: #1319888
Co-Authored-By: Robert Mizielski <robert.mizielski@cloudwatt.com>