Updated openstack/openstack Project: openstack/neutron  b7bf77cc864a4643e11229db055185a09836837c
Stop the dhcp-agent process when dnsmasq version is not determined

When it is not possible to determinate the dnsmasq version
or dnsmasq is not installed, neutron shows only a warning message.
With this commit the dhcp-agent process will be stopped if the
dnsmasq service is not installed.
Modifies the dhcp version unit test with the expected SystemExit exception

Change-Id: Ia26eaae6f432305f412a4b6a4a6b2b7b15d99d45
Closes-Bug: 1212401