Updated openstack/openstack Project: openstack/nova  b16db2bdca8a60d6802adfa34e63d787f5b967d3
Do not allow resize to zero disk flavor

For some reason nova allows creating a flavor with a disk_size of 0,
when an instance is booted using such flavor its disk will use the
native base image size as the size of the ephemeral root volume (from
the ops docs:
http://docs.openstack.org/openstack-ops/content/flavors.html).

Even though having the disk_size=0 means that the instance's disk can
have any imaginable size, nova will still allow to resize that instance
to any flavor defining disk_size >= 0 but it won't shrink the disk if it
is bigger than the new disk_size.

This patch add check to prevent resize to zero disk flavor except
when both current and new disk are 0.

DocImpact

Change-Id: Iffc2f821cc51ccf555ea1e3d7589147375aeb56a
Closes-bug: 1305860