Updated openstack/openstack Project: openstack/neutron  1ed9e7a34236f806ee68ef37a8fb950747934120
Ignore variable column widths in ovsdb functional tests

The ovsdb tests check the output of "ovsdb-client monitor" before and after
some operations. The first line of output is a header row with spaces for
column separation. The column widths can vary depending on the data in the
columns, so compress multiple spaces to one for testing.

Change-Id: I6b4690ea9ef1c1b215411ca2c39b1e65fec2c6c7
Partial-bug: #1336172