Updated openstack/openstack Project: openstack/horizon  6a08a251134011f07685b9e416492fba5e78827a
Added button Add Interface on Network Topology tab

On the Network Topology tab, when a router is selected a new
button was added to allow it connect with different networks.
This change was made following the most simple mechanism that
I could find, please check if it is the correct behavior.

Change-Id: Ia08c5f22e3237fc4e90767a8becfe6413d754b04
Closes-Bug: #1288779