Updated openstack/openstack Project: openstack/tempest  5cfafac5d8a5e0bb1f9657a02f2f7ea61b4f182d
Fix "create a flavor" tests for strong validation

On Nova v2.1/v3 API,  the input validation of "create a flavor" API
has become more strict. As the result it does not allow " 1024 " as
ram size, and does not allow "" as flavor id.

This patch fixes the related tests for passing these tests.

Change-Id: I58a8cc8fd89a12495ed995e06dbd396a82f6ec71
Closes-Bug: #1336173