Updated openstack/openstack Project: openstack/tempest  ff72f4562838fb17458d8a8ac450711793d23372
Add a hacking check to block scenario manager in api tests

This commit adds a new hacking check to ensure that the manager class
from the scenario tests, OfficialClientManager, isn't used in the API
tests. There is already a hacking rule preventing imports of the
official clients, T102. But, the official clients could still be used
through the scenario tests which wouldn't be caught by T102. This
attempts to close that gap by adding a new hacking check for
OfficialClientManager.

Change-Id: I608a1319c136118947e3df74fa88f7124011caa4