Updated openstack/openstack Project: openstack/oslo-incubator  d4d91af23bec67e2df4f4cd10ddb1f33d535b3a8
Remove `processutils` dependency on `log`

Currently, the processutils module depends on openstack.common.log. This
patch proposes to remove this dependency and use the standard library
logging module instead.

The downside of this patch is that it removes the dependency on the log
module but adds a new dependency on the strutils module because of the
`mask_password` call.

Change-Id: I245750f9f397a231891514f6ed1ea9626513cbbb