Updated openstack/openstack Project: openstack/nova  e9a315e671ba3ceaefed52da898293ce6743fb88
Add __repr__ handler for NovaObjects

This adds a __repr__ implementation for the base NovaObject which
should help debugging by compactly describing the contents of an
object. This adds stringification helpers to the Field object(s),
which lets us control what each type looks in compact format. Since
this is just for debugging, we avoid recursing too deeply on nested
objects and the like.

Change-Id: Ic604acfda7e7c7adb913479355726e16ad375f1d