Updated openstack/openstack Project: openstack/horizon  956c99ef5a8e4f52240ee78e8768c26d179e5acb
Use SCSS instead of LESS

Convert all of Horizon's styles from LESS to SCSS, and use pyscss and
django_pyscss to compile them.

This lets us continue with the upgrade to Bootstrap 3, and the use of
django_pyscss also lets us include horizon's style files, with all the
variables, in plugins style files.

This partially implements the blueprint bootstrap-update.

Change-Id: I1bc9b4ded5de2393c8e57e5286b1da373673789d
Closes-bug: #1260675