Updated openstack/openstack Project: openstack/horizon  d79f0f9c65d27a4c4d6b9fa44244c4484c936190
Specify a Network During Database Instance Create

We need to provide a way for users to specify a network during
Database Instance Creation.

Without this, if we're using neutron, the database instance that
comes up does not have a network attached, and consequently the
user doesn't have a way to connect to the Database instance.

Closes-bug: 1331766
Change-Id: I51b9e8fa94196cc6c7fd3b20334f40c690dc216d