Updated openstack/openstack Project: openstack-infra/devstack-gate  a17aeb699abe4f7c7db444f180a1b0583e7b3f57
Handle proposed/ branches in test-matrix.py

We are now cutting proposed/* branches on projects. Test matrix needs to
know about them. Test proposed/foo* as if it were a stable/foo branch
when we know what foo is. If we do not know about foo fallback onto
testing with a master setup as this branch is an intermediate non
integrated release (eg swift proposed/2.0.0) which is released while
other projects are still developing on master.

As part of this change update features.yaml to teach it about juno in
prep for the upcoming juno release.

Change-Id: I3f72d3171d0a1afc00298b1761c06b77bb282a6b