Updated openstack/openstack Project: openstack/oslo-incubator  8b5bc62fe944a07243d86a7ca1a440349a826480
Allow test_service to run in isolation

While trying to port oslo to python 3, we discovered that test_service
wasn't working when not run with testr. SIGHUP tests expect processes to
be a daemon, which is not the case when run directly in process. The fix
can be tested running:

$ python -m openstack.common.lockutils python -m testtools.run \
  tests.unit.test_service

Change-Id: Ie45813668e658789cc133387fce7cd643ec42bb1