Updated openstack/openstack Project: openstack/glance  ea72206627fb75797981d504f8a08c35addfe7e6
Use (# of CPUs) glance workers by default

The config docs have historically recommended that the number of
glance  workers should be set to the number of CPUs available for
best performance, so we should make this the default.

Commit 75c96a48fc7e5dfb59d8258142b01422f81b0253 did the same thing in
Nova in Icehouse and the plan is to do the same thing for Cinder
and Trove workers.

The config files are updated to match the help string in the code.

There is no upgrade impact since glance-api.conf previously
hard-coded the workers value to 1 so anyone upgrading to this
will still get whatever value was set in glance-api.conf prior
to this change.

DocImpact: glance workers will now be equal to the number of
           CPUs available by default if not explicitly specified
           in glance-api.conf and/or glance-registry.conf.

UpgradeImpact: There is no upgrade impact to glance-api workers
           since glance-api.conf previously hard-coded the workers
           value to 1 so anyone upgrading to tihs will still get
           whatever value was set in glance-api.conf prior to
           this change. There is an upgrade impact to the
           glance-registry workers since glance-registry.conf
           did not hard-code the workers value to 1 before change
           I0cee0d284eef9ce5dcb26720499f2c4d31eaca0f, which is
           overwritten here. So anyone upgrading to this change
           that does not have workers specified in
           glance-registry.conf will now be running multiple
           workers by default when they restart the glance
           registry service.

Closes-Bug: #1333325

Change-Id: I6795c6e22268bb3fb67331edc7af641aefa904cc