Updated openstack/openstack Project: openstack/python-keystoneclient  83bef7473ddf39b68c09ae074c741652fc52185d
Add the 'auth' interface type

There are certain requests that will always want to be sent to the
auth_url.

Add a new interface type to the get_endpoint command of the base
identity plugin such that if you ask for the 'auth' interface it will
give you the auth_url.

Implements: blueprint session-auth-endpoint
Change-Id: If653970354b919fdd6e80c061611c3aad129c574